Extract base class for kafka streams tests
