+++ instrumentation/kafka-streams-0.11/javaagent/src/test/groovy/KafkaStreamsSuppressReceiveSpansTest.groovy
@@ -3,90 +3,27 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+
 import io.opentelemetry.api.trace.Span
 import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator
 import io.opentelemetry.context.Context
 import io.opentelemetry.context.propagation.TextMapGetter
-import io.opentelemetry.instrumentation.test.AgentInstrumentationSpecification
 import io.opentelemetry.sdk.trace.data.SpanData
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes
-import org.apache.kafka.clients.admin.AdminClient
-import org.apache.kafka.clients.admin.NewTopic
-import org.apache.kafka.clients.consumer.Consumer
-import org.apache.kafka.clients.consumer.KafkaConsumer
-import org.apache.kafka.clients.producer.KafkaProducer
-import org.apache.kafka.clients.producer.Producer
 import org.apache.kafka.clients.producer.ProducerRecord
-import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.header.Headers
-import org.apache.kafka.common.serialization.IntegerDeserializer
-import org.apache.kafka.common.serialization.IntegerSerializer
 import org.apache.kafka.common.serialization.Serdes
-import org.apache.kafka.common.serialization.StringDeserializer
-import org.apache.kafka.common.serialization.StringSerializer
 import org.apache.kafka.streams.KafkaStreams
 import org.apache.kafka.streams.StreamsConfig
 import org.apache.kafka.streams.kstream.KStream
 import org.apache.kafka.streams.kstream.ValueMapper
-import org.testcontainers.containers.KafkaContainer
-import spock.lang.Shared
 
 import java.time.Duration
-import java.util.concurrent.TimeUnit
 
 import static io.opentelemetry.api.trace.SpanKind.CONSUMER
 import static io.opentelemetry.api.trace.SpanKind.PRODUCER
 
-class KafkaStreamsSuppressReceiveSpansTest extends AgentInstrumentationSpecification {
-
-  static final STREAM_PENDING = "test.pending"
-  static final STREAM_PROCESSED = "test.processed"
-
-  @Shared
-  static KafkaContainer kafka
-  @Shared
-  static Producer<Integer, String> producer
-  @Shared
-  static Consumer<Integer, String> consumer
-
-
-  def setupSpec() {
-    kafka = new KafkaContainer()
-    kafka.start()
-
-    // create test topic
-    AdminClient.create(["bootstrap.servers": kafka.bootstrapServers]).withCloseable { admin ->
-      admin.createTopics([
-        new NewTopic(STREAM_PENDING, 1, (short) 1),
-        new NewTopic(STREAM_PROCESSED, 1, (short) 1),
-      ]).all().get(10, TimeUnit.SECONDS)
-    }
-
-    producer = new KafkaProducer<>(producerProps(kafka.bootstrapServers))
-
-    // values copied from spring's KafkaTestUtils
-    def consumerProps = [
-      "bootstrap.servers"      : kafka.bootstrapServers,
-      "group.id"               : "test",
-      "enable.auto.commit"     : "false",
-      "auto.commit.interval.ms": "10",
-      "session.timeout.ms"     : "60000",
-      "key.deserializer"       : IntegerDeserializer,
-      "value.deserializer"     : StringDeserializer
-    ]
-    consumer = new KafkaConsumer<>(consumerProps)
-
-    // assign topic partitions
-    consumer.assign([
-      new TopicPartition(STREAM_PROCESSED, 0)
-    ])
-  }
-
-  def cleanupSpec() {
-    consumer?.close()
-    producer?.close()
-    kafka.stop()
-  }
+class KafkaStreamsSuppressReceiveSpansTest extends KafkaStreamsBaseTest {
 
   def "test kafka produce and consume with streams in-between"() {
     setup:
@@ -231,18 +168,5 @@ class KafkaStreamsSuppressReceiveSpansTest extends AgentInstrumentationSpecifica
     def spanContext = Span.fromContext(context).getSpanContext()
     spanContext.traceId == streamSendSpan.traceId
     spanContext.spanId == streamSendSpan.spanId
-  }
-
-  private static Map<String, Object> producerProps(String servers) {
-    // values copied from spring's KafkaTestUtils
-    return [
-      "bootstrap.servers": servers,
-      "retries"          : 0,
-      "batch.size"       : "16384",
-      "linger.ms"        : 1,
-      "buffer.memory"    : "33554432",
-      "key.serializer"   : IntegerSerializer,
-      "value.serializer" : StringSerializer
-    ]
   }
 }
